+++ graphql/subcourse/fields.ts
@@ -1,4 +1,5 @@
 import { Prisma } from '@prisma/client';
+import { canPublish } from '../../common/courses/states';
 import { Arg, Authorized, Ctx, Field, FieldResolver, Int, ObjectType, Query, Resolver, Root } from 'type-graphql';
 import { canJoinSubcourse, isParticipant } from '../../common/courses/participants';
 import { CourseState } from '../../common/entity/Course';
@@ -276,5 +277,11 @@ export class ExtendedFieldsSubcourseResolver {
     async canJoin(@Ctx() context: GraphQLContext, @Root() subcourse: Required<Subcourse>, @Arg('pupilId', { nullable: true }) pupilId: number) {
         const pupil = await getSessionPupil(context, pupilId);
         return await canJoinSubcourse(subcourse, pupil);
+    }
+
+    @FieldResolver((returns) => Decision)
+    @Authorized(Role.ADMIN, Role.OWNER)
+    async canPublish(@Root() subcourse: Required<Subcourse>) {
+        return await canPublish(subcourse);
     }
 }
